docs: add page for `usePagination`
