+++ packages/react/src/utilities/hooks/index.ts
@@ -4,6 +4,8 @@ export { useIsomorphicLayoutEffect } from './useIsomorphicLayoutEffect/useIsomor
 export { useMediaQuery } from './useMediaQuery/useMediaQuery';
 export { useRadioGroup } from './useRadioGroup/useRadioGroup';
 export { useSynchronizedAnimation } from './useSynchronizedAnimation/useSynchronizedAnimation';
+export { usePagination } from './usePagination/usePagination';
 
 export type { UseRadioGroupProps } from './useRadioGroup/useRadioGroup';
 export type { UseCheckboxGroupProps } from './useCheckboxGroup/useCheckboxGroup';
+export type { UsePaginationProps } from './usePagination/usePagination';
